https://github.com/opencart/opencart/issues/10987

+++ upload/system/library/log.php
@@ -38,6 +38,6 @@ class Log {
      *
      */
 	public function __destruct() {
-		file_put_contents($this->file, $this->message, FILE_APPEND);
+		if (is_file($this->file)) file_put_contents($this->file, $this->message, FILE_APPEND);
 	}
 }
